RAG/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
asyncpg>=0.29.0
httpx>=0.27.0
qdrant-client>=1.10.0
langchain-text-splitters>=0.3.0
pydantic-settings>=2.4.0
